+++ include/display.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "scale.hpp"
+#include "config.hpp"
 #include <SPI.h>
 #include <U8g2lib.h>
